// src/types/wager.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@0.25.2.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export type Uint128 = string
export interface InstantiateMsg {
  amounts: Uint128[]
  collection_address: string
  expiries: number[]
  fairburn_bps: number
  fee_address: string
  fee_bps: number
  matchmaking_expiry: number
  max_currencies: number
}
export type ExecuteMsg =
  | {
      update_config: {
        params: ParamInfo
      }
    }
  | {
      set_winner: {
        current_prices: [Decimal, Decimal]
        prev_prices: [Decimal, Decimal]
        wager_key: [number, number]
      }
    }
  | {
      wager: {
        against_currencies: Currency[]
        currency: Currency
        expiry: number
        token: number
      }
    }
  | {
      cancel: {
        token: number
      }
    }
export type Decimal = string
export type Currency =
  | 'dot'
  | 'avax'
  | 'uni'
  | 'atom'
  | 'link'
  | 'near'
  | 'icp'
  | 'sand'
  | 'btc'
  | 'eth'
  | 'bnb'
  | 'xrp'
  | 'ada'
  | 'doge'
  | 'sol'
  | 'mana'
  | 'cake'
  | 'ar'
  | 'osmo'
  | 'rune'
  | 'luna'
  | 'ustc'
  | 'stars'
  | 'mir'
export const currencies = [
  'dot',
  'avax',
  'uni',
  'atom',
  'link',
  'near',
  'icp',
  'sand',
  'btc',
  'eth',
  'bnb',
  'xrp',
  'ada',
  'doge',
  'sol',
  'mana',
  'cake',
  'ar',
  'osmo',
  'rune',
  'luna',
  'ustc',
  'stars',
  'mir',
]
export interface ParamInfo {
  amounts?: Uint128[] | null
  collection_address?: string | null
  expiries?: number[] | null
  fairburn_bps?: number | null
  fee_address?: string | null
  fee_bps?: number | null
  matchmaking_expiry?: number | null
  max_currencies?: number | null
}
export type QueryMsg =
  | {
      wagers: {}
    }
  | {
      wager: {
        token: number
      }
    }
  | {
      matchmaking: {}
    }
  | {
      token_status: {
        token: number
      }
    }
  | {
      config: {}
    }
export type Addr = string
export interface ConfigResponse {
  config: Config
}
export interface Config {
  amounts: Uint128[]
  collection_address: Addr
  expiries: number[]
  fairburn_percent: Decimal
  fee_address: Addr
  fee_percent: Decimal
  matchmaking_expiry: number
  max_currencies: number
}
export type Timestamp = Uint64
export type Uint64 = string
export interface MatchmakingResponse {
  matchmaking: MatchmakingItemExport[]
}
export interface MatchmakingItemExport {
  against_currencies: Currency[]
  amount: Uint128
  currency: Currency
  expires_at: Timestamp
  expiry: number
  token: NFT
}
export interface NFT {
  collection: Addr
  token_id: number
}
export type TokenStatus =
  | 'none'
  | {
      matchmaking: MatchmakingItemExport
    }
  | {
      wager: WagerExport
    }
export interface TokenStatusResponse {
  token_status: TokenStatus
}
export interface WagerExport {
  amount: Uint128
  expires_at: Timestamp
  wagers: [WagerInfo, WagerInfo]
}
export interface WagerInfo {
  currency: Currency
  token: NFT
}
export interface WagerResponse {
  wager: WagerExport
}
export interface WagersResponse {
  wagers: WagerExport[]
}